+++ CVI/CapaVista/Mantenimientos/frmDepartamento.cs
@@ -67,6 +67,19 @@ namespace CapaVista.Mantenimientos
             rbDeshabilitado.Checked = false;
         }
 
-        
+        private void txtdeparnombre_TextChanged(object sender, EventArgs e)
+        {
+            txtEstado.Text = "1";
+        }
+
+        private void rbHabilitado_CheckedChanged(object sender, EventArgs e)
+        {
+            txtEstado.Text = "1";
+        }
+
+        private void rbDeshabilitado_CheckedChanged(object sender, EventArgs e)
+        {
+            txtEstado.Text = "0";
+        }
     }
 }
